requirements.txt
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.9
clean-text==0.5.0
click==8.0.3
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0x`
dataclasses==0.8
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.2.10
django-cors-headers==3.10.1
djangorestframework==3.13.1
emoji==1.6.1
filelock==3.4.0
ftfy==6.0.3
httplib2==0.9.2
huggingface-hub==0.2.1
idna==3.3
importlib-metadata==4.8.3
install==1.3.5
joblib==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
packaging==21.3
pandas==1.1.5
pexpect==4.2.1
Pillow==5.1.0
praw==7.5.0
prawcore==2.3.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.0.6
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2021.3
pyxdg==0.25
PyYAML==6.0
regex==2021.11.10
reportlab==3.4.0
requests==2.26.0
requests-unixsocket==0.1.5
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.5.4
SecretStorage==2.3.1
simplejson==3.13.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
system-service==0.3
systemd-python==234
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.1
tqdm==4.62.3
transformers==4.14.1
typing-extensions==4.0.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
update-checker==0.18.0
urllib3==1.26.7
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.2.5
websocket-client==1.2.3
xkit==0.0.0
zipp==3.6.0
zope.interface==4.3.2
Brlapi==0.6.6